# src/main/resources/static/api.yml
openapi: 3.0.0
info:
  title: Grid Dynamics Internship -  Hospital Management
  description: |-
    A simple application demo for hospital management created at the Grid Dynamics Java internship. The app uses Java 17, Spring boot, Gradle, Hibernate, Postgres + database migration tool (Liquibase), Lombok, Swagger.
  version: 0.0.1
servers:
  - url: https://localhost:8080
tags:
  - name: Patients
    description: Available commands relatied to patients
  - name: Prescriptions
    description: Available commands related to prescriptions
  - name: Events
    description: Available commands related to events
  - name: Treatments
    description: Available commands related to treatments
  - name: Admin
    description: Available commands for admin / auth
paths:
  /treatments:
    get:
      tags:
        - Treatments
      summary: Get all treatments
      description: ''
      operationId: getAllTreatments
      responses:
        '200':
          description: All treatments successfully fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TreatmentResponse'
    post:
      tags:
        - Treatments
      summary: Create treatment with given JSON
      description: ''
      operationId: addTreatment
      requestBody:
        description: Treatment JSON for new entry
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTreatment'
      responses:
        '200':
          description: Treatment info successfully fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatmentResponse'
  /treatments/{treatmentID}:
    get:
      tags:
        - Treatments
      summary: Get treatment info for treatment with provided ID
      description: ''
      operationId: getTreatmentById
      parameters:
        - name: treatmentID
          description: ID of the given treatment
          in: path
          required: true
          schema:
            type: string
            example: UUID
      responses:
        '200':
          description: Treatment info successfully fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatmentResponse'
    put:
      tags:
        - Treatments
      summary: Modify treatment with provided ID, replace with given JSON
      description: ''
      operationId: modifyTreatment
      parameters:
        - name: treatmentID
          description: ID of the given treatment
          in: path
          required: true
          schema:
            type: string
            example: UUID
      requestBody:
        description: Treatment JSON that will replace the current entry
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTreatment'
      responses:
        '200':
          description: Treatment info successfully fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatmentResponse'
    delete:
      tags:
        - Treatments
      summary: Delete treatment with provided ID
      description: ''
      operationId: deleteTreatment
      parameters:
        - name: treatmentID
          description: ID of the given treatment
          in: path
          required: true
          schema:
            type: string
            example: UUID
      responses:
        '200':
          description: Treatment successfuly deleted
  /patients:
    get:
      tags:
        - Patients
      summary: Get all patient files
      description: ''
      operationId: getAllPatients
      responses:
        '200':
          description: Patient files successfully fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientResponse'
    post:
      tags:
        - Patients
      summary: Add a patient
      description: Adds provided patient to database
      operationId: addPatient
      requestBody:
        description: Patient info to be added to the database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPatient'
      responses:
        '200':
          description: Patient successfully added to database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
  /patients/{patientId}:
    get:
      tags:
        - Patients
      summary: Get patient file with the given ID
      description: ''
      operationId: getPatient
      parameters:
        - name: patientId
          in: path
          description: Unique patient identifier from Patients table
          required: true
          schema:
            type: string
            example: UUID
      responses:
        '200':
          description: Patient file successfully fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
    put:
      tags:
        - Patients
      summary: Edit patient file of the given ID
      description: ''
      operationId: editPatient
      parameters:
        - name: patientId
          in: path
          description: ID of the patient that will be edited
          required: true
          schema:
            type: string
            example: UUID
      requestBody:
        description: Patient's JSON file that will replace existing one
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPatient'

      responses:
        '200':
          description: patient successfully modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
  /patients/{patientId}/discharge:
    put:
      tags:
        - Patients
      summary: Discharge a patient by given ID
      description: Deletes any prescriptions under the given insurance number, subsequently all related events.
      operationId: dischargePatientById
      parameters:
        - name: patientId
          in: path
          description: The patient's unique identifier found on the Patients table.
          required: true
          schema:
            type: string
            example: UUID
      responses:
        '200':
          description: Successfully discharged patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
  /patients/{patientId}/events:
    get:
      tags:
        - Events
      summary: Get all events for a patient with given ID
      description: ''
      operationId: getAllEventsFor
      parameters:
        - name: patientId
          in: path
          required: true
          description: Unique patient identifier in patients table
          schema:
            type: string
            example: UUID
      responses:
        '200':
          description: All events successfully fetched for patient ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponse'
  /patients/{patientId}/prescriptions:
    get:
      tags:
        - Prescriptions
      summary: Get prescriptions for the patient with the given ID
      description: ''
      operationId: getPrescriptionsOf
      parameters:
        - name: patientId
          in: path
          description: Unique patient identifier from Patients table
          required: true
          schema:
            type: string
            example: UUID
      responses:
        '200':
          description: Successfully fetched prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrescriptionResponse'
    post:
      tags:
        - Prescriptions
      summary: Prescribe procedures or medication
      description: Prescription is for the patient
      operationId: createPrescription
      parameters:
        - name: patientId
          in: path
          description: UUID of patient that will be prescribed
          required: true
          schema:
            type: string
            example: UUID
      requestBody:
        description: Created prescription object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPrescription'
      responses:
        '200':
          description: Successfully added prescription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionResponse'
  /prescriptions/{prescriptionId}:
    get:
      tags:
        - Prescriptions
      summary: Get prescription info of given prescription ID
      description: ''
      operationId: getPrescription
      parameters:
        - name: prescriptionId
          in: path
          description: Unique prescription identifier in Prescriptions table
          required: true
          schema:
            type: string
            example: UUID
      responses:
        '200':
          description: Prescription successfully fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionResponse'
    put:
      tags:
        - Prescriptions
      summary: Modify a given prescription
      description: Edit prescription (appointment) by ID, prescription parameter and its value
      operationId: modifyPrescription
      parameters:
        - name: prescriptionId
          in: path
          description: Unique id of the prescription that will be modified
          required: true
          schema:
            type: string
            example: UUID
      requestBody:
        description: Prescription JSON file that will replace the one with the provided ID.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPrescription'
      responses:
        '200':
          description: Prescription successfully modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionResponse'
    delete:
      tags:
        - Prescriptions
      summary: Delete the given ID's prescription
      description: ''
      operationId: deletePrescriptionById
      parameters:
        - name: prescriptionId
          in: path
          description: The given prescription's ID in Prescriptions table
          required: true
          schema:
            type: string
            example: UUID
      responses:
        '200':
          description: Prescription successfully removed
  /events/{eventId}:
    get:
      tags:
        - Events
      summary: Get event info for event with given ID
      description: ''
      operationId: getEventById
      parameters:
        - name: eventId
          in: path
          description: Unique event ID of the event that will be modified
          required: true
          schema:
            type: string
            example: UUID
      responses:
        '200':
          description: Operation Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
    put:
      tags:
        - Events
      summary: Modify an event's parameters
      description: ''
      operationId: modEvent
      parameters:
        - name: eventId
          in: path
          description: Unique event ID of the event that will be modified
          required: true
          schema:
            type: string
            example: UUID
      requestBody:
        description: Event JSON file that will replace the entry at given Event ID
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEvent'
      responses:
        '200':
          description: Operation Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
  /events/{eventId}/cancel:
    put:
      tags:
        - Events
      summary: Cancel given event
      description: ''
      operationId: cancelEvent
      parameters:
        - name: eventId
          in: path
          description: ID of the event that will be canceled
          required: true
          schema:
            type: string
            example: UUID
        - name: cancelReason
          in: query
          description: Reason for cancelling the event
          required: true
          schema:
            type: string
            example: Doctor's orders
            default: Not Specified
      responses:
        '200':
          description: Successfully canceled event
  /users:
    post:
      tags:
        - Admin
      summary: Create new user
      description: ''
      operationId: createUser
      requestBody:
        description: User Object that will be passed for creation
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '200':
          description: Successfully created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /auth/login:
    post:
      tags:
        - Admin
      summary: Authenticate with given username and password
      description: ''
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Auth token in response header -> "authorization:"
  /auth/refresh:
    post:
      tags:
        - Admin
      summary: Process provided refresh token
      description: ''
      operationId: tokenRefresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JwtRefreshRequest'
      responses:
        '200':
          description: 'Refresh token succesfully processed, returning JWT auth token and refresh token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
components:
  schemas:
    PatientResponse:
      type: object
      properties:
        id:
          type: string
          example: UUID
        name:
          type: string
          description: Name of the patient
          example: John Smith
        diagnosis:
          type: string
          description: Patient's diagnosis
          example: Tuberculosis
        insuranceNumber:
          type: string
          example: AO8323
        doctorId:
          type: string
          description: UUID
        status:
          type: string
          description: patient's treatment progress
          example: In treatment
          enum:
            - In treatment
            - recovered
    NewPatient:
      type: object
      properties:
        name:
          type: string
          description: Name of the patient
          example: John Smith
        diagnosis:
          type: string
          description: Patient's diagnosis
          example: Tuberculosis
        insuranceNumber:
          type: string
          example: "A0123331"
        doctorId:
          type: string
          example: UUID
          description: doctor's id in doctor catalogue
        status:
          type: string
          description: patient's treatment progress
          example: In treatment
          enum:
            - In treatment
            - recovered
    NewPrescription:
      type: object
      properties:
        treatmentId:
          type: string
          example: UUID
          description: The prescribed treatment's unique id in catalogue.
        timePattern:
          type: string
          description: Frequency of treatment
          example: Once a day
        period:
          type: integer
          format: int64
          example: 5
          description: Duration of the treatment in days
    PrescriptionResponse:
      type: object
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        patientId:
          type: string
          example: 123e4547-e89b-12d3-a456-426614174000
          description: The patient's id in patients catalogue.
        treatmentId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174233
          description: The prescribed treatment's unique id in catalogue.
        timePattern:
          type: string
          description: Frequency of treatment
          example: Once a day
        period:
          type: integer
          format: int64
          example: 5
          description: Duration of the treatment in days
        datePrescribed:
          type: string
          format: date-time
          description: Date and time of prescription
    NewEvent:
      type: object
      properties:
        prescriptionId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174223
          description: Related prescription's id in Prescriptions table
        dateTime:
          type: string
          format: date-time
          description: Scheduled date and time for the event.
        status:
          type: string
          description: Current event status
          example: Done
          enum:
            - Scheduled
            - Done
    EventResponse:
      type: object
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174233
          description: Unique event id
        prescriptionId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174223
          description: Related prescription's id in Prescriptions table
        patientId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174342
          description: Related patent's id in Patient table
        dateTime:
          type: string
          format: date-time
          description: Scheduled date and time for the event.
        status:
          type: string
          description: Current event status
          example: Cancelled
          enum:
            - Scheduled
            - Done
            - Cancelled
        cancelReason:
          type: string
          description: Reason for cancelling the event
          example: Doctor's orders
          default: n/a
    NewTreatment:
      type: object
      properties:
        name:
          type: string
          example: Colonoscopy
          description: Name of procedure/medicine
        type:
          type: string
          description: Whether the treatment is a procedure or a medicine
          example: Procedure
          enum:
            - Procedure
            - Medicine
    TreatmentResponse:
      type: object
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-4266141742334
          description: Unique treatment identifier in catalogue
        name:
          type: string
          example: Colonoscopy
          description: Name of procedure/medicine
        type:
          type: string
          description: Whether the treatment is a procedure or a medicine
          example: Procedure
          enum:
            - Procedure
            - Medicine
    NewUser:
      type: object
      properties:
        firstName:
          type: string
          example: John
          description: The user's first name
        lastName:
          type: string
          example: Doe
          description: The user's last name
        userName:
          type: string
          example: aolvera
          description: The user's login identifier
        password:
          type: string
          format: password
          example: ao12!B45
          description: The user's password
        role:
          type: string
          example: "admin"
          description: User's authorization level
          enum:
            - admin
            - doctor
            - nurse
    UserResponse:
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174233
          description: User's unique identifier in Users table
        firstName:
          type: string
          example: John
          description: The user's first name
        lastName:
          type: string
          example: Doe
          description: The user's last name
        userName:
          type: string
          example: dBowie
          description: Username
        password:
          type: string
          format: password
          example: ao12!B45
          description: The user's password
        createdAt:
          type: string
          format: date-time
          description: Date and time in which the User was created
        createdBy:
          type: string
          example: "drDooLittle"
          description: Username of this user's creator
        role:
          type: string
          example: "doctor"
          description: User's authorization level
          enum:
            - admin
            - doctor
            - nurse
          x-enum-varnames: [ROLE_ADMIN, ROLE_DOCTOR, ROLE_NURSE]
    Credentials:
      properties:
        username:
          type: string
          example: jPerez
          description: The user's ID
        password:
          type: string
          example: ASsda233F2!
          description: The user's password
    JwtRefreshRequest:
      properties:
        refreshToken:
          type: string
          example: "9fea96ad-57b3-4635-b5f2-06719e0af0a4"
          description: The refresh token that will be refreshed
    JwtResponse:
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhZG1pbiIsImV4cCI6MTY3MTEzNTQ5NCwibGFzdE5hbWUiOiJTdXBlciBVc2VyIn0.arMRwjwFoIS5s_kNPc1EwA_p5niYSLXCZLCZ4hy5k70"
          description: The JWT auth token that was refreshed
        refreshToken:
          type: string
          example: "9fea96ad-57b3-4635-b5f2-06719e0af0a4"
          description: The refresh token associated with the JWT token
  requestBodies:
    NewPatient:
      description: Patient object that will be added to the database
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NewPatient'